+++ packages/test-utils/src/loadEmojiData.ts
@@ -3,12 +3,10 @@ import { Emoji } from 'emojibase';
 let DATA: Emoji[] = [];
 
 export default function loadEmojiData(): Emoji[] {
-  if (DATA.length > 0) {
-    return DATA;
+  if (DATA.length === 0) {
+    // eslint-disable-next-line global-require, @typescript-eslint/no-unsafe-assignment
+    DATA = require('../test-data.json');
   }
-
-  // eslint-disable-next-line global-require, @typescript-eslint/no-unsafe-assignment
-  DATA = require('../test-data.json');
 
   return DATA;
 }